sconi
This is a project that I launched in a forum (manew.com).

Finally, the final day of the end of the month finished the general content. The time after work is constantly perfected. There are still a lot of shortcomings, and some details are not dealt with.

To achieve all the basic functions of the props, the bomb 1 generation of all the animations of the monster, the barrier can be unlimited, temporarily set to 10. Map monsters are automatically generated.

In order to facilitate customs clearance, the walls of the props are deliberately set in different colors.

At the end, there are colored eggs, but the premise is to play the tenth pass, saying that I did not play!

The source code is arranged and released.